pyproject.toml
[project]
name = "hotpot-retrieval-qa"
version = "0.1.0"
description = "RAG system for multi-hop question answering using HotpotQA dataset"
authors = [
    {name = "Zak Raicik"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.15"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
    [tool.poetry.dependencies]
datasets = "^4.0.0"
    sentence-transformers = "^5.1.0"
    faiss-cpu = "^1.12.0"
